backend/requirements.txt
fastapi>=0.110.0
uvicorn==0.24.0
pymongo>=4.6.0
dnspython>=2.4.0
python-multipart==0.0.6
pydantic>=2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-dotenv>=1.0.0
groq>=0.4.0

# LangChain AI Agent
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=1.0.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langsmith>=0.2.0
Pillow>=10.0.0
cloudinary>=1.36.0

